requirements.txt
django==5.1.4
gunicorn==23.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
